+++ spec/sources/circle_spec.rb
@@ -38,4 +38,16 @@ describe Danger::CISource::CircleCI do
     expect(t.base_commit).to eql("759adcbd0d8f")
     expect(t.head_commit).to eql("13c4dc8bb61d")
   end
+
+  it 'handles funky compare URLs' do
+    env = {
+      "CIRCLE_BUILD_NUM" => "true",
+      "CI_PULL_REQUEST" => "https://github.com/artsy/eigen/pull/800",
+      "CIRCLE_COMPARE_URL" => "https://github.com/artsy/eigen/compare/759adcbd0d8f^...13c4dc8bb61d"
+    }
+    t = Danger::CISource::CircleCI.new(env)
+    expect(t.base_commit).to eql("759adcbd0d8f^")
+    expect(t.head_commit).to eql("13c4dc8bb61d")
+  end
+
 end